Converted raw pointers to smart pointers.

+++ src/ECS/ECS.cpp
@@ -104,7 +104,7 @@ Entity Registry::CreateEntity()
     Entity entity(entityId);
     pendingCreationEntities.insert(entity);
 
-    if (entityId >= entityComponentSignatures.size())
+    if (entityId >= static_cast<int>(entityComponentSignatures.size()))
     {
         entityComponentSignatures.resize(entityId + 1);
     }
